\name{bdeR}
\alias{bdeR}
\title{bdeR - a library in R to access IMB data.}
\usage{
getVariables()
getLocation()
getData(codVar='1;2')
}
\description{
This is a module to access data from IMB (Mauro Borges Statistic and Socioeconomic Institute).'.
}
\examples{
getVariables()
getLocation()
getData(codVar='1;2')
}
